# platform/openapi.yaml
openapi: 3.1.0
info:
  title: clickbus-iris
  version: 1.0.0
  description: Clickbus Iris platform API — authentication, organization management, microservice tokens, and metrics ingestion.
  license:
    name: Apache-2.0
servers:
  - url: https://iris.clickbus.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie set after successful sign-in.
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Internal server error
...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: jane@example.com
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          format: uuid
        password:
          type: string
          format: password
          minLength: 8
          description: Must contain at least one uppercase letter, one lowercase letter, one number, and one special character.
...
    Complete2FARequest:
      type: object
      required:
        - userId
        - code
        - method
      properties:
        userId:
          type: string
          format: uuid
        code:
          type: string
          minLength: 6
          maxLength: 6
          description: 6-digit verification code.
          example: "123456"
        method:
          type: string
          enum:
            - totp
            - email
    Complete2FAResponse:
      type: object
      required:
        - success
        - loginToken
        - user
      properties:
        success:
          type: boolean
          example: true
        loginToken:
          type: string
          format: uuid
          description: Temporary login token valid for 5 minutes.
        user:
          type: object
          required:
            - id
            - email
            - name
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            name:
              type: string
            image:
              type:
                - string
                - "null"
...
    Resend2FARequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          format: uuid
    UserOrganization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type:
            - string
            - "null"
        logo_url:
          type:
            - string
            - "null"
        role:
          type: string
          enum:
            - owner
            - admin
            - member
    GetUserOrgsResponse:
      type: object
      required:
        - organizations
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/UserOrganization"
...
  responses:
    BadRequest:
      description: Bad request - validation error or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or session invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Request a password reset link
      description: |
        Generates a password reset token and sends an email. Always returns 200 to avoid leaking whether the email exists.
        Rate limit tier: strictAuthRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset link sent (or no-op if email not found).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: If an account with that email exists, we have sent a password reset link.
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Reset Password
  # ──────────────────────────────────────────────
  /api/auth/reset-password:
    post:
      operationId: resetPassword
      summary: Reset password with token
      description: |
        Resets the user password using a previously issued reset token. Validates password strength.
        Rate limit tier: strictAuthRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Password reset successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Verify Email
  # ──────────────────────────────────────────────
...
  /api/auth/complete-2fa:
    post:
      operationId: complete2FA
      summary: Complete 2FA during login
      description: |
        Validates a 2FA code (TOTP or email OTP) during the login flow and returns a temporary login token.
        Supports TOTP backup codes as a fallback when method is "totp".
        Locks the account after 5 consecutive failed attempts for 15 minutes.
        Rate limit tier: authRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Complete2FARequest"
      responses:
        "200":
          description: 2FA completed; temporary login token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complete2FAResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Account locked due to too many failed attempts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Resend Verification Email
  # ──────────────────────────────────────────────
...
  /api/auth/resend-2fa:
    post:
      operationId: resend2FA
      summary: Resend 2FA email OTP code
      description: |
        Generates and sends a new 2FA email OTP code to the user.
        Rate limit tier: strictAuthRateLimit.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Resend2FARequest"
      responses:
        "200":
          description: Verification code sent successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Verification code sent successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Get User Organizations
  # ──────────────────────────────────────────────
  /api/auth/get-user-orgs:
    get:
      operationId: getUserOrgs
      summary: List organizations for the current user
      description: |
        Returns all active organizations the authenticated user belongs to, including role and logo.
        Also triggers auto-accept domain membership checks.
        Rate limit tier: apiRateLimit.
      tags:
        - auth
      security:
        - sessionCookie: []
      responses:
        "200":
          description: List of user organizations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserOrgsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Auth - Providers
  # ──────────────────────────────────────────────
...